Use Program2.cs and GraphNode2.cs which partly fix mutual relationships.
The way the seed code is written suggests that the student should
not have to worry about ensuring that appropriate reverse edges
are added when loading data. The existing code would add the
necessary edges for spouse, friend, and parent edges but would
not add the reverse parent edge when the input defines a
child.