Fix provided friends command by reconcile inconsistency.
All of the sample .rg files list relationship edges using a format
like hasRelationship. E.g., hasFriend and hasParent. The code in
Program.cs which creates edges took these edge names verbatim. But the
code in Program.cs which recognizes mutual relationships and adds
appropriate reverse edges recognizes edges with names like friend and
parent. To reconcile this, now the code will only treat a record
property as being relationship-defining if it begins with "has" and
extrants and converts the remaining string from PascalCase to
camelCase. This causes hasFriend to create an edge in the model called
friend. Now loading plan.rg and listing the friends of Case functions.

This change reconciles an inconsistency bug within the seed code. It
fixes functionality that the student be able to assume works out of
the box.