Improve C# style by using properties various changes.
The main goal of this change is to convert Java-style getter methods
to C# properties. GraphEdge and GraphNode now expose properties like
Name and Label as properties rather than methods. RelationshipGraph
exposes its list of all nodes through the IReadOnlyList<T> interface
rather than allowing public updating (this, this is different from
using immutable types which are "secure" against casting, but this is
a sort of good faith measure and users of the API may not assume that
a cast to a writable type will work because the method of
implementation is not guaranteed).

The documentation of much of GraphEdge and GraphNode was updated to
use VS/sandcastle-style documentation tags.

I also could not help but to update some code that accessed the
changed properties to use linq methods or C#-6.0 features.