Use MEF to dynamically enumerate commands.
Now to define a command, just [Export(typeof(Command))]. It is
recommended when creating a new command to just use an existing
command as a template anyway.

No need to edit Program.cs! Now the student merely creates a class per
command listed in the project rubric. All gradable code is neatly
constrainted to a small set of well-defined files which contain
minimal boilerplate!