Provide helpful message for anticipated Command creation mistake.
If one wants to create a new Command, they might use an existing
Command class as a template. If that person fails to update the base()
call with the name of the new command, the program would fail with an
unhelpful message about duplicate dictionary keys (I think). This
change explicitly checks for that mistake and tries to provide a
helpful error message describing how to resolve it.